# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ICT Helpdesk & Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List all tickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Submit a new ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Not found
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/comments:
    get:
      operationId: listTicketComments
      tags: [tickets]
      summary: List comments on a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketComment"
    post:
      operationId: addTicketComment
      tags: [tickets]
      summary: Add a comment to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketCommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketComment"

  # ── Assets ───────────────────────────────────────────────────────────────
...
  /maintenance:
    get:
      operationId: listMaintenanceSchedules
      tags: [maintenance]
      summary: List maintenance schedules
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Maintenance schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceSchedule"
    post:
      operationId: createMaintenanceSchedule
      tags: [maintenance]
      summary: Create a maintenance schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceSchedule"
  /maintenance/{id}:
    get:
      operationId: getMaintenanceSchedule
      tags: [maintenance]
      summary: Get maintenance schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceSchedule"
    patch:
      operationId: updateMaintenanceSchedule
      tags: [maintenance]
      summary: Update maintenance schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceScheduleUpdate"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceSchedule"
    delete:
      operationId: deleteMaintenanceSchedule
      tags: [maintenance]
      summary: Delete maintenance schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Knowledge Base ───────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all staff users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Ticket:
      type: object
      required:
        - id
        - title
        - category
        - subcategory
        - priority
        - status
        - reportedBy
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: "hardware | software | network | user_support"
        subcategory:
          type: string
        priority:
          type: string
          description: "low | medium | high | critical"
        status:
          type: string
          description: "open | in_progress | on_hold | resolved | closed"
        reportedBy:
          type: string
        reporterEmail:
          type: ["string", "null"]
        reporterDepartment:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        assetId:
          type: ["integer", "null"]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
    TicketInput:
      type: object
      required:
        - title
        - category
        - subcategory
        - priority
        - reportedBy
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        priority:
          type: string
        reportedBy:
          type: string
        reporterEmail:
          type: string
        reporterDepartment:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assetId:
          type: ["integer", "null"]
    TicketUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        priority:
          type: string
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        resolution:
          type: string
    TicketComment:
      type: object
      required:
        - id
        - ticketId
        - authorName
        - content
        - createdAt
      properties:
        id:
          type: integer
        ticketId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        isInternal:
          type: boolean
        createdAt:
          type: string
    TicketCommentInput:
      type: object
      required:
        - authorName
        - content
      properties:
        authorName:
          type: string
        content:
          type: string
        isInternal:
          type: boolean

    # ── Asset schemas ──────────────────────────────────────────────────────
...
    MaintenanceSchedule:
      type: object
      required:
        - id
        - title
        - assetCategory
        - frequency
        - scheduledDate
        - status
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        assetCategory:
          type: string
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        frequency:
          type: string
          description: "daily | weekly | monthly | quarterly | annually | one_time"
        scheduledDate:
          type: string
        completedDate:
          type: ["string", "null"]
        status:
          type: string
          description: "scheduled | in_progress | completed | overdue | cancelled"
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    MaintenanceScheduleInput:
      type: object
      required:
        - title
        - assetCategory
        - frequency
        - scheduledDate
      properties:
        title:
          type: string
        description:
          type: string
        assetCategory:
          type: string
        assetId:
          type: ["integer", "null"]
        frequency:
          type: string
        scheduledDate:
          type: string
        assignedTo:
          type: ["integer", "null"]
        notes:
          type: string
    MaintenanceScheduleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assetCategory:
          type: string
        assetId:
          type: ["integer", "null"]
        frequency:
          type: string
        scheduledDate:
          type: string
        completedDate:
          type: string
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        notes:
          type: string

    # ── Knowledge Base schemas ─────────────────────────────────────────────
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - department
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          description: "admin | ict_officer | user"
        department:
          type: string
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required:
        - name
        - email
        - role
        - department
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        phone:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        phone:
          type: string
        isActive:
          type: boolean

    # ── Dashboard schemas ──────────────────────────────────────────────────
...