# lib/api-zod/src/generated/types/ticketComment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ICT Helpdesk & Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketComment {
  id: number;
  ticketId: number;
  authorName: string;
  content: string;
  isInternal?: boolean;
  createdAt: string;
}