# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ICT Helpdesk & Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  /** admin | ict_officer | user */
  role: string;
  department: string;
  /** @nullable */
  phone?: string | null;
  isActive?: boolean;
  createdAt: string;
}